Move Export to separate component
