+++ frontend/src/components/Settings.js
@@ -1,9 +1,9 @@
 import React from "react";
-import Form from "react-bootstrap/Form";
 import Button from "react-bootstrap/Button";
 import Container from "react-bootstrap/Container";
 import { level } from "../helpers/Level";
 import Import from "./Import";
+import Export from "./Export";
 
 export default class Settings extends React.Component {
   constructor(props) {
@@ -12,22 +12,17 @@ export default class Settings extends React.Component {
   }
 
   render() {
-    const data = JSON.stringify({
-      programs: this.props.programs,
-      pkgs: this.props.pkgs,
-      groups: this.props.groups,
-      rules: this.props.rules,
-      ranges: this.props.ranges,
-      users: this.props.users,
-    });
-
     return (
       <>
         <Container fluid>
-          <Form.Group>
-            <Form.Label>Exportovaná data:</Form.Label>
-            <Form.Control as="textarea" value={data} readOnly />
-          </Form.Group>
+          <Export
+            programs={this.props.programs}
+            pkgs={this.props.pkgs}
+            groups={this.props.groups}
+            rules={this.props.rules}
+            ranges={this.props.ranges}
+            users={this.props.users}
+          />
           {this.props.userLevel >= level.ADMIN && (
             <Import
               programs={this.props.programs}
@@ -39,12 +34,10 @@ export default class Settings extends React.Component {
               client={this.props.client}
             />
           )}
-        </Container>
-        {this.props.userLevel >= level.ADMIN && (
-          <Container fluid>
+          {this.props.userLevel >= level.ADMIN && (
             <Button onClick={this.deleteAll}>Smazat vše</Button>
-          </Container>
-        )}
+          )}
+        </Container>
       </>
     );
   }
